Fix for the overflow issue while processing GPS based timestamp from RU
The absslot_tx overflow issue observed during integration testing of AW2S
RU with GPS as the source for the clock synchronization. Causes the gNB
not to start the TX thread and stops the processing.

+++ executables/nr-gnb.c
@@ -115,10 +115,10 @@ static void tx_func(void *param)
   int slot_tx = info->slot;
   int frame_rx = info->frame_rx;
   int slot_rx = info->slot_rx;
-  int absslot_tx = info->timestamp_tx / info->gNB->frame_parms.get_samples_per_slot(slot_tx, &info->gNB->frame_parms);
-  int absslot_rx = absslot_tx - info->gNB->RU_list[0]->sl_ahead;
+  int64_t absslot_tx = info->timestamp_tx / info->gNB->frame_parms.get_samples_per_slot(slot_tx, &info->gNB->frame_parms);
+  int64_t absslot_rx = absslot_tx - info->gNB->RU_list[0]->sl_ahead;
   if (absslot_rx < 0) {
-    LOG_W(NR_PHY, "Slot ahead %d is larger than absslot_tx %d. Cannot start TX yet.\n", info->gNB->RU_list[0]->sl_ahead, absslot_tx);
+    LOG_W(NR_PHY, "Slot ahead %d is larger than absslot_tx %ld. Cannot start TX yet.\n", info->gNB->RU_list[0]->sl_ahead, absslot_tx);
     return;
   }
   LOG_D(NR_PHY, "%d.%d running tx_func\n", frame_tx, slot_tx);
@@ -218,8 +218,8 @@ void rx_func(void *param)
   for (; i < gNB->frame_parms.slots_per_subframe / 2; i++)
     cumul_samples += gNB->frame_parms.get_samples_per_slot(i, &gNB->frame_parms);
   int samples = cumul_samples / i;
-  int absslot_tx = info->timestamp_tx / samples;
-  int absslot_rx = absslot_tx - gNB->RU_list[0]->sl_ahead;
+  int64_t absslot_tx = info->timestamp_tx / samples;
+  int64_t absslot_rx = absslot_tx - gNB->RU_list[0]->sl_ahead;
   int rt_prof_idx = absslot_rx % RT_PROF_DEPTH;
   clock_gettime(CLOCK_MONOTONIC,&info->gNB->rt_L1_profiling.start_L1_RX[rt_prof_idx]);
   start_meas(&softmodem_stats_rxtx_sf);
